move common code to common.go
