+++ cmd/common.go
@@ -349,3 +349,34 @@ func stringToBytes(fieldAttr ReinterpretField, value string) []byte {
 	}
 	return buf
 }
+
+func newSchemaTree(reader *reader.ParquetReader) *schemaNode {
+	schemas := reader.SchemaHandler.SchemaElements
+	stack := []*schemaNode{}
+	root := &schemaNode{
+		SchemaElement: *schemas[0],
+		Children:      []*schemaNode{},
+	}
+	stack = append(stack, root)
+
+	pos := 1
+	for len(stack) > 0 {
+		node := stack[len(stack)-1]
+		if len(node.Children) < int(node.SchemaElement.GetNumChildren()) {
+			childNode := &schemaNode{
+				SchemaElement: *schemas[pos],
+				Children:      []*schemaNode{},
+			}
+			node.Children = append(node.Children, childNode)
+			stack = append(stack, childNode)
+			pos++
+		} else {
+			stack = stack[:len(stack)-1]
+			if len(node.Children) == 0 {
+				node.Children = nil
+			}
+		}
+	}
+
+	return root
+}
